+++ history/__init__.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from .backends import get_backend, session  # noqa
 from .utils import get_history_model  # noqa
 
-__version__ = "3.4.1"
+__version__ = "3.4.2"
 __version_info__ = tuple(
     int(num) if num.isdigit() else num
     for num in re.findall(r"([a-z\d]+)", __version__, re.I)